feat(dispatcher/routing): core route related classes added/moved
