+++ prototype/dispatch/delivery-dispatch/core/routing/src/main/java/dev/aws/proto/core/routing/route/IRouter.java
@@ -14,9 +14,9 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.aws.proto.core.routing.interfaces;
+package dev.aws.proto.core.routing.route;
 
-import dev.aws.proto.core.routing.Coordinate;
+import dev.aws.proto.core.routing.location.Coordinate;
 
 import java.util.List;
