+++ prototype/dispatch/delivery-dispatch/core/routing/src/main/java/dev/aws/proto/core/routing/route/SegmentRoute.java
@@ -1,5 +1,6 @@
-package dev.aws.proto.core.routing;
+package dev.aws.proto.core.routing.route;
 
+import dev.aws.proto.core.routing.distance.Distance;
 import lombok.Data;
 
 @Data
